Update patch set 1
Patch Set 1:

I think it would be nice with an api/ subdirectory for the things related to ice. I'm afraid I'm not familiar with what the code under p2p/ really is doing.

For the things belonging to the call:rtp_interfaces target, that should maybe be moved into api/rtp.

Patch-set: 1
CC: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>