Update patch set 2
Patch Set 2:

Sorry for the slow response Magnus, this required me digging back a bit to this CL: https://codereview.webrtc.org/2854123003/

The method CreateModularPeerConnectionFactory() is called with a fully created mediaengine as an argument. The reason for that is to allow building the lib with only data channel support. Naming didn't turn out the best here - one idea with the "modular" in there was that this would be the 'root' create method and called by all the others. That's only half true.

I suggest we move forward removing all unused arguments from CMPCF() to avoid confusion, video codec factories as well - they should work the same as audio in this case. Feel free to come up with a better name as well.

Patch-set: 2
Reviewer: Gerrit User 5077 <5077@58829da1-049c-39fb-b951-ebdcd0984767>