Update patch set 7
Patch Set 7:

(3 comments)

Can you please ask taylor or someone in his team who probably is more up to date with the API work and what the next step is.
peerconnectioninterface.h is in my mind the API. So if we move it out from there, it should be temporary.

yes, candidate.h is probably an API.

Patch-set: 7
Reviewer: Gerrit User 5578 <5578@58829da1-049c-39fb-b951-ebdcd0984767>