Update patch set 1
Patch Set 1:

> May have missed something. But basically, the way our code is designed, we have a lot of stacks of transport objects. When an object at a higher level needs to know about something at a lower level, the signal has to propagate all the way to the top. sigslot::repeater would make this much simpler. Also, see my recent email to the webrtc core group.

If you change to virtual methods, then one alternative is to let the SignalBaz() method of the higher level object return a pointer (or reference) directly to the right slot in a lower level object. Then we get forwarding logic at the signal setup time, rather than when signals are invoked, which seems preferable to me.

I hope the boilerplate for SignalBaz in the forwarding class shouldn't be worse than for setting up a sigslot repeater.

I see there might be a problems if for some reason we want to tear down the higher level object long before the lower level object; then users of the higher level object don't automatically get disconnected from the lower-level object. sigslot repeater makes it easier to break up the connection in this case. If that's going to cause any a real problems, I don't know.

Patch-set: 1
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>