Update patch set 1
Patch Set 1: Code-Review+1

Can you add documentation in VideoFrame.I420Buffer that the implementations have to return a slice where the markers can be modified by the caller?

lgtm after addressing this

Can you break the commit message on words and not in the middle of a word? Also, the new syntax for marking a bug is "Bug: " instead of "BUG=".

Patch-set: 1
Reviewer: Gerrit User 5112 <5112@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1