Update patch set 11
Patch Set 11:

(2 comments)

For context, this change was suggested during review of https://webrtc-review.googlesource.com/c/src/+/2684

I think it's good to not expose more reference count details than we really need to. I had a quick look at the docs for IUNKNOWN, and it says that return values from AddRef and Release are "intended to be used only for test purposes", so I don't think there's much value in return value consistency with IUNKNOWN. And the main feature of IUNKNOWN is QueryInterface (a bit like C++ dynamic_cast), which isn't relevant here.

I also had a quick look at how std::shared_ptr exposes the ref_count. It has a use_count method documented like this:

  Common use cases include:

  comparison with ​0​. [...]

  comparison with 1. If use_count returns 1, there are no other owners. (The 
  deprecated member function unique() is provided for this use case.) In 
  multithreaded environment, this does not imply that the object is safe to 
  modify because accesses to the managed object by former shared owners may 
  not have completed, [...]

In our case, we require that objects are modified only if HasOneRef is true.

The common case would be a thread gaining exclusive access to an object (e.g., by creating it), modify it, and then share references with other threads. When the other thread accesses the object (for reading, or for HasOneRef + modify), we need to guarantee that other threads see all of the modification done by the first thread. I hope our use AtomicOps when reading or modifying the refcount are strong enough, but if so, why can't std::shared_ptr promise the same?

(I'm not familiar with the fine details of the C++ memory model).

Patch-set: 11
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0