Update patch set 5
Patch Set 5:

Sorry for the flood of messages. Accidentally clicked submit.
Will repeat my message so it's not lost.

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > Is there any good reason to remove this? If the files are already there, it'll just do a hash check.
> > 
> > Please don't submit yet.
> 
> The reason is because we won't download the files if we run this test on swarming. They should be already available, and we should just pack them and send them to the bot.

Yes, you can pack them and send them to the bot. This code does not affect the ability to do that, so why remove it? If the files are already there it's just going to check the hash.

  0> File tools_webrtc/audio_quality/linux/pesq exists and SHA1 matches. Skipping.
  2> File tools_webrtc/audio_quality/linux/PolqaOem64 exists and SHA1 matches. Skipping.

And you get that check for free, without the need to check for file existence manually like in this change.

Also note that the same is being done in another place:
https://cs.chromium.org/chromium/src/third_party/webrtc/examples/androidtests/video_quality_loopback_test.py?l=105

Patch-set: 5
Reviewer: Gerrit User 5114 <5114@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0