Update patch set 12
Patch Set 12:

> Patch Set 11:
> 
> On Tue, Oct 3, 2017, 09:25 Karl Wiberg (Gerrit) <> wrote:

My email reply didn't make it...here it is again:

Ok, I would have kept it in the interface but don't have a strong preference. Is it to cut down on typing or do we want to allow usage elsewhere?

Patch-set: 12
Reviewer: Gerrit User 5508 <5508@58829da1-049c-39fb-b951-ebdcd0984767>