Update patch set 1
Patch Set 1:

> Where that is needed, the way I suggest won't work. B needs a record of connected callbacks, which is the point of the sigslot::repeater.

With that in mind, is this ok to land this? Or should another reviewer take a look? Your suggestion is a good idea, by the way; it's more efficient and less code to maintain in the situations where it works.

Patch-set: 1
Reviewer: Gerrit User 5053 <5053@58829da1-049c-39fb-b951-ebdcd0984767>