Update patch set 5
Patch Set 5:

(2 comments)

Turns out that adding video_frame_api to the api breaks nacl (!) in exciting ways. The error was extremely difficult to understand:

../../third_party/webrtc/system_wrappers/source/aligned_malloc.cc -o newlib_pnacl/obj/third_party/webrtc/system_wrappers/system_wrappers/aligned_malloc.o
../../third_party/webrtc/system_wrappers/source/aligned_malloc.cc:13:10: fatal error: 'memory.h' file not found
#include <memory.h>

Like, um, wat. The explanation is that newlib_pnacl, whatever that is, depends on libjingle_peerconnection_api. However, the special compiler or build environment is allergic to memory.h (and some other stuff in system_wrappers), so pulling in system_wrappers is poisonous. I really dislike this difficult-to-understand dependency. Either way, I removed the dependency on video_frame_api, which was pulling in system_wrappers, and temporarily squelched the warnings that caused me to add it; we'll have to discuss whether video_frame should move into API. I will do that in a follow up CL.

Patch-set: 5
Reviewer: Gerrit User 5027 <5027@58829da1-049c-39fb-b951-ebdcd0984767>