Update patch set 7
Patch Set 7:

Problem with video_frame_api seems to be in i420_buffer.h, declaring a private member

  const std::unique_ptr<uint8_t, AlignedFreeDeleter> data_;

It's unclear to me if the special alignment magic (part of system_wrappers) brings any real benefit. We could purge it from the header file only by not using unique_ptr here and make this hidden in the destructor implementation, but I'm not sure that's the best way to deal with; it won't eliminate the link-time dependency.

Patch-set: 7