Update patch set 2
Patch Set 2:

(4 comments)

I think I may have miscommunicated about the purpose of this class; it's not intended to just own the DTLS and SRTP transport objects, but also be responsible for setting up the SRTP transport based on the result of the DTLS handshake. That way, an object using it (e.g., BaseChannel) doesn't need to deal with any SRTP or DTLS specifics; it can just wait for it to become writable and then start using it. (The encrypted header extensions are one exception, but that should be handled in TransportController/JsepTransport in the future.)

Patch-set: 2
Reviewer: Gerrit User 5053 <5053@58829da1-049c-39fb-b951-ebdcd0984767>