Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > Please provide some details on what runtime this test has and an estimate for commitbot+trybot runs we have per day so Sana can get an idea on the increased load this will have on the touch-swarming pool (and if it can handle it).
> > 
> > Sure.
> > Runtime is about 50 minutes, and we have ~35 runs a day.
> 
> This is actually kind of long, I'm not sure if touch-swarming can support it. +cc sergeyberezin

Indeed. We have 11 iPhone 6s devices, and they are already pretty busy with occasional pending queues:

http://vi/chrome_infra/Jobs/per_job?service_name=touch-swarming&job_regexp=.*device%3AiPhone+6s.*&duration=1d&pool=.*device%3AiPhone8.*&refresh=-1

You seem to have up to 4 builds per hour at peak times: http://shortn/_twtqr0iTJm , which would take ~1/3 of the pool. I wouldn't be surprised if we start timing out tasks (and not just yours, but others' too). The peaks seem to happen at EMEA times, so this may compete with some Chrome/iOS developers in Europe.

IMHO, we *might* be able to handle the load, but barely. And if we start seeing timeouts, we'll have to revert this change.

Note also that you're using iOS 10.3.3 devices. Because iOS 11 is out, we can no longer expand the pool. Consider switching to iOS 11 devices.

Patch-set: 3
CC: Gerrit User 5718 <5718@58829da1-049c-39fb-b951-ebdcd0984767>