Update patch set 6
Patch Set 6:

> Patch Set 3:
> 
> (4 comments)
> 
> Would it be worth making this templetized?

Sorry, forgot to answer this one before.
IMHO, it's not worth templating it as it only works with unsigned int types. And then it requires data to have a light long tail, which makes uint64_t out of the question. 1- and 2-bytes integers are possible, but not so frequent in practice and memory savings by using them instead of int32_t would be only in the map and they are negligible if there is a short tail indeed. Otherwise, just extending a histogram_low to whole range is a nice idea, and then it doesn't matter how many bytes the index in the array is.

Patch-set: 6
Reviewer: Gerrit User 5117 <5117@58829da1-049c-39fb-b951-ebdcd0984767>