Update patch set 6
Patch Set 6:

jiancui: Sorry this CL has been taking a while. Thanks a bunch for your patience!

deadbeef: What are you thoughts w.r.t. public API changes? (I asked them on an earlier patch set but probably it got lost.)

The issue is SSLCertificate is an unfortunate API. It is simultaneously a list of certificates and a single certificate itself. Which means one must answer questions like:

1. Should SSLCertificate::FromPEMString parse just the first certificate as it does today, or should it keep parsing and return a leaf + chain?

2. Should SSLCertificate::ToDER serialize just the one certificate as it does today or concatenate them all?

3. Should SSLCertificate::ToPEMString serialize just the one certificate as it does today or concatenate them all?

For (1), without parsing the whole chain, it becomes pretty difficult to test stuff because there's no way to parse a chain.

For (2) and (3), in the one certificate version, if the caller wants DER serializations of each, they must iterate over cert->GetChain(). For DER in particular, I think this is definitely correct because people don't just randomly concatenate DER encodings. PEM is a bit more common, but then it seems weird for ToDER and ToPEMString to behave different? There's also a question of which changes constitute a backwards compatibility break.

Unfortunately, I don't see another real implementation of SSLCertificate to compare against...

Oh! Actually, we can look back to the NSS days. That one says that, in each of (1), (2) and (3), those APIs only acted on a single certificate.

Here's a proposal then:

1. Keep FromPEMString as it was before, and make FromPEMStringWithChain do the new thing.

2. ToDER is left as before. Callers that want to serialize every cert should just loop over GetChain.

3. ToPEMString is left as before. Callers that want to serialize every cert should just loop over GetChain.

Patch-set: 6
Reviewer: Gerrit User 5602 <5602@58829da1-049c-39fb-b951-ebdcd0984767>