Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > jiancui: Sorry this CL has been taking a while. Thanks a bunch for your patience!
> > 
> > deadbeef: What are you thoughts w.r.t. public API changes? (I asked them on an earlier patch set but probably it got lost.)
> > 
> > The issue is SSLCertificate is an unfortunate API. It is simultaneously a list of certificates and a single certificate itself. Which means one must answer questions like:
> 
> First of all: if SSLCertificate being both a list of certificates and a single certificate is an unfortunate API, why not use SSLCertChain for the list? Is this infeasible for some reason? I may be missing a bit of context on this CL.

SSLCertChain and SSLCertificate are existing WebRTC APIs. They weren't added in this CL. They just weren't implemented in the OpenSSL port before, just the now-deleted NSS port.

Unfortunately, it looks like the way you all have it set up, an SSLCertificate has an SSLCertChain, which has a list of non-leaf SSLCertificates (that promise not to have SSLCertChains).

That means that a method like SSLCertificate::ToPEMString is ambiguous between whether you are talking about one certificate or the associated chain. WebRTC's API does not provide an object which is just the leaf.

> > 1. Should SSLCertificate::FromPEMString parse just the first certificate as it does today, or should it keep parsing and return a leaf + chain?
> > 
> > 2. Should SSLCertificate::ToDER serialize just the one certificate as it does today or concatenate them all?
> > 
> > 3. Should SSLCertificate::ToPEMString serialize just the one certificate as it does today or concatenate them all?
> 
> It would be surprising to me if SSLCertificate::ToPEMString(cert.FromPEMString()) != cert. Same with ToDER/DerToPem/FromPEMString. So, I'd suggest either making it handle the list in each case, or documenting the behavior very clearly in a comment (suggesting iterating over GetChain if you want the whole chain).

Making ToDER handle the list is definitely nuts. That is not a format used by anyone. Remember that the core object in X.509 is a single certificate. Lists of certificates are structures you built on top of it. But WebRTC's API conflates the two levels, thus the problem.

Note also that teaching the other functions to act on lists rather than individual certificates may not be a backwards-compatible change.

Patch-set: 6
Reviewer: Gerrit User 5602 <5602@58829da1-049c-39fb-b951-ebdcd0984767>