Add style guidance about forward declarations.
We prefer the Google style guide over the chromium guide in this case:
avoid forward declarations whenever possible. We don't have the same
compilation time issue that chromium does, so we can afford to do this.

The majority of our code already follows this guideline, as far as I'm
aware, though some forward declarations are still used to resolve
circular dependencies.

Bug: None
Notry: true
Change-Id: I712e0361acf76217067b13b8b3e4bc931d2a0238