Update patch set 1
Patch Set 1:

> > Do you know what the counters.retransmitted thing is ultimately used for
> No, unfortunately I don't, so it might be good idea to check where it is used.
> 
> 
> Did another digging into the past and found CL where IsPacketRetransmitted was formed:
> https://webrtc-codereview.appspot.com/373006

Looks like it is used for rtcp feedback, from CalculateRtcpStatistics in this file.

  // Only for report blocks in RTCP SR and RR.
  last_report_inorder_packets_ =
      receive_counters_.transmitted.packets -
      receive_counters_.retransmitted.packets;

But from the name of this name, it's not clear if it really should do the complex retransmit checks, or just the seqno-based in-order check?

And then it's also used for aplication stats, I think, e.g, StreamDataCounters::MediaPayloadBytes is documented as excluding retransmissions.

I think my main confusion is: What really is the intended distinction between !IsInOrder and IsRetransmitted?

So assume a packet is not "InOrder". Possible reasons:

1. Reorder or duplication done by the best-effort network. 

2. Spontaneous duplicate retransmission from the sender, for padding or as a simple fec.

3. Retransmissions done by the sender in response to NACK.

In which of these cases is IsRetransmitted intended to return true?

Or if it's really duplicate packets we want to exclude in the stats, should we keep a record of recently received seqnos, and say that "IsRetransmitted" means that we've seen that seqno before? That seems to be the right thing at least for the MediaPayloadBytes use case.

Stefan, do you know the intention? You reviewed the introduction of the rtt input to IsRetransmitted logic back in 2012 (see cl Danil dug up).

(And the reason for looking at this is that I was trying to make the receive statistics be simply an RtpPacketSinkInterface, and then we'd need some other way to get access to the rtp module's rtt estimate, if we still think it's needed).

Patch-set: 1
Reviewer: Gerrit User 5525 <5525@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>