Update patch set 22
Patch Set 22:

> Post() takes a raw pointer to the MessageHandler, though, so I am not surprised that there are code paths where it's leaked. My preferred course of action in this sort of case is to make a separate cleanup CL where I change that raw pointer to unique_ptr; if there are no problems, the compiler will help you eliminate all ways it could leak.

It's MessageQueue::Post. It looks like the typical use is to pass |this| for the MessageHandler argument, so it's not likely that we can make it take ownership.

As for OnAudioChannelGone, it looks like the object will be deleted when it gets a call to OnMessage. If the thread is being stopped, OnMessage is never called.

Since RemoteAudioSource::MessageHandler is a just proxy object, I wonder if it would work to just let RemoteAudioSource inherit MessageHandler (as a private inherit), and Post to |this|. Logic should be the same, but we don't need this heap allocation of the proxy.

Taylor, it looks like you simplified this code a while ago, in cl https://codereview.webrtc.org/2046173002. Do you have any advice?

Patch-set: 22
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>