Update patch set 8
Patch Set 8:

(3 comments)

Is there any real disadvantage of storing a vector of things that contain X509s, as opposed to storing a STACK_OF(X509)? If not, this is certainly a lot simpler.

It really would be preferable if this were covered by a test, though. Would it be possible to add some way for the server to use a certificate chain, either by having OpenSSLIdentity support an SSLCertChain or some other means?

Patch-set: 8
Reviewer: Gerrit User 5053 <5053@58829da1-049c-39fb-b951-ebdcd0984767>