Expose network adapter enumeration options in iOS SDK
Expose port allocator flags:
* PORTALLOCATOR_DISABLE_ADAPTER_ENUMERATION
* PORTALLOCATOR_DISABLE_DEFAULT_LOCAL_CANDIDATE
in the iOS SDK.

The bug 8432 is caused by trying to connect through a
"link-local" interface (IP address 169.254.0.x/16),
which is listed among the iPhone network interfaces.
The bug is not happening after setting the exposed flags
inside the RTCConfiguration,
which is passed to the RTCPeerConnectionFactory.

Bug: webrtc:8432
Change-Id: Ibc3aaf462fef8c98ae159e5fc32b06f8fad1820e