Update patch set 1
Patch Set 1:

> I understand the efforts to modularize codecs that we discussed, however this project doesn't necessarily fit either of the internal/external categories.
It feels like there is still confusion around this. Soon, there will be no built-in "internal" codecs. The current "internal" codecs (VP8, VP9, and H264) and the wrappers around them (Simulcast, VideoEncoderSoftwareFallbackWrapper) will be available in the codebase, but it's up to the client to put them together and send them into PeerConnectionFactory. And "external" does not mean that the code does not reside in WebRTC. You can more about it here: go/injectable_sw_codecs 

> - We are not planning to do it only on Chromium and bake an extra external codec into RTCVideoEncoder. We want use it even in standalone WebRTC applications. Landing it a behind a flag inside WebRTC is the goal. 
I understand it will not only be used by Chromium, but it can't be added by default to all clients. You can land the code in WebRTC, but it will still be up to the client to pass the stereo codec into PeerConnectionFactory. We don't need to use a flag for this.

> - Yet, we cannot keep it within InternalEncoderFactory because we want to use HW/external encoder if available.
> - Overall, we want to use HW/external encoder when available and fall back to SW/internal encoder otherwise. It is exactly the Simulcast use case, wrapping an external or internal implementation.
> - I followed how Simulcast gets created in WebRtcVideoEngine. Keeping Simulcast and Stereo adapter at the same level makes more sense to me. Let me know if there are plans to remove simulcast out of this class too.
The mapping external->HW, internal->SW is not accurate anymore, all codecs will be "external". Yes, Simulcast is on its way out of WebRtcVideoEngine as well.

> - Additionally, there are SDP parts where things still need to be Stereo codec specific, i.e. see SelectStereoAssociatedVideoCodec() logic. Just like RTX, we refer to other payload types from Stereo codec. Adding another generic SDP line wouldn't carry the info needed.
Yes, I think this will be the tricky part and we need to come up with a solution.

> - An option I can think for modularity is to add |StereoAdapterExternalEncoderFactory(cricket::WebRtcVideoEncoderFactory* external_encoder_factory)| that wraps the external factory(if --enable-stereo-codec command line flag is available) such that it exposes all the external supported codecs as they are, and adds stereo to supported list(might be just referring to internal or external encoder underneath). It will still need the SDP logic built in WebRtcVideoEngine though, and it would be very different than Simulcast although doing similar job. WDYT?
I don't think we need this. I think it will be clearer once I make the CL for splitting out Simulcast from WebRtcVideoEngine. I'm currently blocked on https://webrtc-review.googlesource.com/c/src/+/13980, but it will land soon.

> - Either way, the small CL up for review is doing a bit that isn't really related to these. Can you comment on the appropriateness of the folder it lives in? Currently bots fail due to a gn dependency loop because "media/engine/" can't be referenced from "modules/video_coding/codecs/". Should I move the folder to "media/engine/" or somewhere completely new?
> https://webrtc-review.googlesource.com/c/src/+/11841
I like where it's placed, although I'm not an owner of that folder. You can avoid the dependency if you use the new webrtc::VideoCodecFactory that resides in the api folder instead. I commented in your small CL.

Patch-set: 1