Update patch set 1
Patch Set 1:

> It feels like there is still confusion around this. Soon, there will be no built-in "internal" codecs. The current "internal" codecs (VP8, VP9, and H264) and the wrappers around them (Simulcast, VideoEncoderSoftwareFallbackWrapper) will be available in the codebase, but it's up to the client to put them together and send them into PeerConnectionFactory. And "external" does not mean that the code does not reside in WebRTC. You can more about it here: go/injectable_sw_codecs 

Thanks for the link, great detail. I dropped a couple comments. I hope those can be helpful as I faced these when adding a new codec which is exactly the use case. 

> I understand it will not only be used by Chromium, but it can't be added by default to all clients. You can land the code in WebRTC, but it will still be up to the client to pass the stereo codec into PeerConnectionFactory. We don't need to use a flag for this.

Yes, that is the plan.

> The mapping external->HW, internal->SW is not accurate anymore, all codecs will be "external". Yes, Simulcast is on its way out of WebRtcVideoEngine as well.

Well, that was the most confusing part for me, but it makes things very clear if it is moving out. 

> I don't think we need this. I think it will be clearer once I make the CL for splitting out Simulcast from WebRtcVideoEngine. I'm currently blocked on https://webrtc-review.googlesource.com/c/src/+/13980, but it will land soon.

Looking forward to it. As I said, this POC CL is for discovering codebase, testing and prototyping. We can land it by adjusting to the new interfaces at the same time. I don't see a definite blocker.

> I like where it's placed, although I'm not an owner of that folder. You can avoid the dependency if you use the new webrtc::VideoCodecFactory that resides in the api folder instead. I commented in your small CL.

Sure, I will continue the discussion there.

Patch-set: 1
Reviewer: Gerrit User 5682 <5682@58829da1-049c-39fb-b951-ebdcd0984767>