Remove content_name from BaseChannel constructor and add id
This is part of the work to make BaseChannel an RtpTransceiver.

An RtpTransceiver does not always have a mid set (e.g., if it's
created by a call to addTransceiver before setting SDP). To allow
the RtpTransceiver to be created without an associated session
description, this change renames content_name -> mid and no longer
requires it in the constructor. Instead, it will be set by a new
function set_mid.

A place where content_name was used before SDP might be set is
with the stats collection code. To address this, a new id has been
added that is a random UUID for the purpose of identification and
stats.

Bug: webrtc:7600
Change-Id: I2b1cb3faa8cc0cb0de8f82509b95256ece6a8658