Replace BaseChannel Init_w with SetTransports
This is part of the work to make BaseChannel into an RtpTransceiver.

An RtpTransceiver can be re-used, so it does not make sense to treat
the first time a session description is set any different than later
SetLocalDescription/SetRemoteDescription calls.

This bulk of Init_w was to call SetTransports anyways. The other
initialization was:
- Setting the RtcpMuxFilter active if RTCP multiplexing was required.
    This has been moved to the constructor.
- Setting the BaseChannel as the media channel's NetworkInterface.
    This has been moved to SetTransports.
- Hooking up signals for the RtpDataChannel. This has been moved to
    the RtpDataChannel constructor.

Bug: webrtc:7600
Change-Id: I4cee7ed42137dbec7b9ec5f25a3fff16ed3c02a6