Adding //third_party/auto to WebRTC DEPS file.
Starting from https://chromium-review.googlesource.com/750645, 
Chromium started to depend on //third-party/auto in order to use
an errorprone plugin that disallows the usage of synchronized public
methods.

This CL adds the dependency to //third_party/auto and adds some
suppression in order to avoid to break the build if synchronized
public methods are used.

Bug: webrtc:8491
Change-Id: Ie9bb70520fc713dc294050c8a536ce5091b8339c