Update patch set 1
Patch Set 1:

(1 comment)

I'm a little surprised this is causing an issue. The description shouldn't be destroyed if the SetRemoteDescription helper function succeeds (it should be set as the remote description). And if it doesn't succeed, we just return anyway without touching |desc|. I also don't see anything in the follow-up CL that looks like it would change this behavior, so I'm curious how this surfaced?

Also, would you be willing to do this for SetLocalDescription too?

Patch-set: 1
Reviewer: Gerrit User 5039 <5039@58829da1-049c-39fb-b951-ebdcd0984767>