Update patch set 2
Patch Set 2:

FYI, we hit a crash from the original CL downstream, so thank you for reverting :)

Before the original change, if no internal decoder could be created for a type (e.g. H.264), you'd get a NullVideoDecoder. That can happen when there's an external decoder factory but it returns nullptr when asked to create a specific decoder.

In the new implementation, it'll return nullptr, which will (at least) trigger this CHECK:
11-08 16:13:41.249  8659  8988 E rtc     : 
11-08 16:13:41.249  8659  8988 E rtc     : 
11-08 16:13:41.249  8659  8988 E rtc     : #
11-08 16:13:41.249  8659  8988 E rtc     : # Fatal error in third_party/webrtc/files/stable/webrtc/video/video_receive_stream.cc, line 119
11-08 16:13:41.249  8659  8988 E rtc     : # last system error: 11
11-08 16:13:41.249  8659  8988 E rtc     : # Check failed: decoder.decoder
11-08 16:13:41.249  8659  8988 E rtc     : # 
11-08 16:13:41.249  8659  8988 E rtc     : #

Patch-set: 2
CC: Gerrit User 5501 <5501@58829da1-049c-39fb-b951-ebdcd0984767>
CC: Gerrit User 6013 <6013@58829da1-049c-39fb-b951-ebdcd0984767>