Avoid use-after-move in PeerConnection::SetRemoteDescription.
In SetRemoteDescription we have a raw pointer |desc| whose ownership is
passed to a helper function. Before this CL we continue to use |desc|
after ownership is passed under the assumption that the object is not
deleted.

In this CL, we instead rely on remote_description() after the helper
function has been called. In practice, this is a pointer to the same
object, but it removed the assumption about |desc| being valid after
its ownership is passed.

Bug: None
Change-Id: I144a190ea00f303f4713b64c45aa3e811c0f4b2e