Removing conditional visibility.
Conditional visibility is complex to maintain and it is not well supported by other build systems.

This CL removes it and falls back on the more relaxed visibility value ("*" in this case).
It is not a problem because the targets that were using conditional visibility are all marked as testonly and this is probably enough to keep the build graph clean.

Bug: None
Change-Id: I2d2b5ac9a02d08c2863950116db455976ee1459c