Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> modules/video_coding lgtm, but I think asapersson@ should have a look at the fallback wrapper changes to ensure that they are OK with the forced fallback.

Yes, please review that. One thing I changed is that the initial assignment of |forced_fallback_possible_| does not use 'PayloadStringToCodecType(codec.name) == kVideoCodecVP8' (because I want to remove the codec argument). I looked at the rest of the code and it doesn't seem necessary to have this check, because we will check if the codec is VP8 again when InitEncode is called.

Patch-set: 2
Reviewer: Gerrit User 5098 <5098@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0