Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> How is this different than issue https://bugs.chromium.org/p/webrtc/issues/detail?id=6749, and the fix in https://codereview.webrtc.org/2798993002?

That patch doesn't fix all the possible cases because layoutSubviews is not always called when the orientation changes.  Some examples:
- if you start the app in portrait mode and rotate it directly to upside down (without passing through landscape) layoutSubviews is not called and the video is shown face down.  
- if you attach a different capture session with the RTCCameraPreviewView.setCaptureSession at some point after the UIView has been created layoutViews is not called again and the orientation of the video is not updated.

In addition in my patch I enable rotation in your example apps so that this issue can be tested and noticed if broken again in the future.

Does it make sense?

Patch-set: 2
Reviewer: Gerrit User 6000 <6000@58829da1-049c-39fb-b951-ebdcd0984767>