Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (1 comment)
> 
> When using raw pointers there's always the life time aspect to consider. I don't know the bitrate controller code well enough to be certain this will not lead to bugs.

The change should not be able to introduce a bug of that kind, as the previous implementation also had a raw pointer hidden within the object whose raw pointer was returned by the create function. Note how it also was up to the caller to actually destroy the interim object, for instance by storing it in a unique_ptr.

From that perspective I'd argue that it's an improvement to show this fact in a more explicit manner by using a direct raw pointer. There might be further improvements to be made by using other pointer types. Although for better or worse, throwing around raw pointers seems to be a common pattern in the code base.

Patch-set: 4
CC: Gerrit User 5532 <5532@58829da1-049c-39fb-b951-ebdcd0984767>