Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> Hi Sami,
> 
> This will break when errorprone is updated and will continue to break as new warnings are introduced. Is there a specific reason you want javac warnings to become errors? The intention of warnings is to avoid breaking the build.
> 
> Thanks,
> 
> Peter

Hi,

Previously we had a lot of warnings and it was very difficult to notice when new ones were introduced accidentally. It was also very hard to notice this as a reviewer. By making them errors, it is very clear when new warnings are introduced.

It should relatively straightforward to suppress warnings when they are introduced by new rolls and assign bug to fix them to the code author / me.

Patch-set: 11