Update patch set 2
Patch Set 2:

First attempt to rewrite measurements for cpu adaptation. Unclear if we still need the OnTargetFramerateUpdated logic, or anything other special to handle irregular frame rate.

Currently, if an encoded frame arrives after a long interval, with encode time x and interval d much larger than the time constant tau (5s), the load estimate will be set to x / d which is going to be pretty close to 0. We probably don't want to signal underuse due to this, but I'm not sure if we'll do that or not with the current adaptation logic.

Patch-set: 2
Reviewer: Gerrit User 5524 <5524@58829da1-049c-39fb-b951-ebdcd0984767>
CC: Gerrit User 5525 <5525@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false