Update patch set 6
Patch Set 6: Commit-Queue+2

> Patch Set 6: Code-Review+1
> 
> I'm still not sure this is the right way to go, but I might be wrong, and we could probably fix it later if I'm right.  So I'll not block this CL any further.

Sure, if this turns out to be the wrong approach we can fix it later on.

I think a list of "remote streams" that is independent of receivers (what we have today) should go away. IMO, remote_streams() should be the result of iterating receivers to find streams, just like getReceivers() is the result of iterating transceivers to find receivers. This is most like the spec, where the receivers have the [[AssociatedRemoteMediaStreams]] slot. Nothing else needs to have a streams slot. Our code should reflect that even on the lower layers.

Patch-set: 6
Reviewer: Gerrit User 5142 <5142@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=+2